Added Objects, Added Collision, fixed minor bugs in C Edition
Added new featue objects, these objects right now don't do a lot but in the future updates objects will be able to interact with their environment and other objects too, which will definetly open lots possibilities only limited by creativity.

Objects may not necessarily be visible on the canvas, but they can be operated in the background.

Collision has been added to objects, which detect if anything is present in the surroundings of the object.

------------------------------------------

Fixed a bug in insert_x_line and insert_y_line where if the y2 value exceeds canvas_height but y1 dosent, the line is not created.